+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.1.1 - ?
 
 - Manage the groups of a netbird account with the `netbird::Group` resource
+- Manage the setup keys of a netbird account with the `netbird::SetupKey` resource
 
 ## v0.1.0 - 2026-08-11